requirements.txt
# Audio Analysis 
openai-whisper==20230314 # 또는 faster-whisper
ffmpeg-python==0.2.0  # 영상에서 음성 추출
librosa
pydub==0.25.1 # 오디오 포맷 변환
soundfile
Flask==3.0.3                   # API 엔드포인트 생성을 위한 웹 프레임워크
gunicorn==22.0.0               # 프로덕션 환경에서 Flask 앱을 실행하기 위한 WSGI 서버

# Deep Learning 
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
timm==1.0.15

# NLP / Transformers 
transformers==4.35.0  # GPT / BERT류 모델 로딩
sentence-transformers==2.2.2 # 문장 임베딩 후 유사도 분석
scikit-learn==1.3.2   # 간단한 ML

# Data & Visualization 
numpy==1.24.4
pandas==2.2.2
matplotlib==3.8.2
seaborn==0.13.2
tqdm==4.66.2

# 기타 
click==8.1.7
Pillow==9.4.0
python-dotenv==1.0.1  # 환경 변수 관리
scipy

# API Clients 
google.generativeai==0.8.5